crud message

+++ base/views.py
@@ -1,24 +1,97 @@
 from django.shortcuts import render, redirect
-from .models import Room
+from django.db.models import Q
+from django.contrib.auth.models import User
+from .models import Room, Topic, Message
 from .form import RoomForm
-
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 # Create your views here.
 
 
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User dose not exits.')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password does not exit.')
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    context = {'form': form}
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred while registering')
+
+    return render(request, 'base/login_register.html', context)
+
+
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms': rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) | Q(name__icontains=q) | Q(description__icontains=q))
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
 
     return render(request, 'base/home.html', context)
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body'),
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
+    context = {'room': room, 'room_messages': room_messages,
+               'participants': participants}
 
     return render(request, 'base/room.html', context)
 
 
+@ login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -32,9 +105,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@ login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed to update this room')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -48,13 +125,33 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@ login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed to update this room')
 
     if request.method == 'POST':
         room.delete()
         return redirect("home")
 
     context = {'obj': room}
+
+    return render(request, 'base/delete.html', context)
+
+
+@ login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('Your are not allowed to update this room')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect("home")
+
+    context = {'obj': message}
 
     return render(request, 'base/delete.html', context)
